Remove input schemas check from jsonBodyHandler
This behaviour was ported from the old version of
osprey-method-handler. The check limited body schema
types which were validated in jsonBodyHandler to
Scalar, Node, Array only.

The issue was discovered when trying to validate a
body of type Union. The validation just did not happen
because the body schema was ignored as it didn't pass
the check.

Now the check is completely removed as it doesn't seem
to make sense. As jsonBodyHandler is only called for
application/json requests, all the request bodies
defined in RAML for application/json should work well
with validation.